rpserver/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
django-cors-headers
psycopg2-binary